chore(plugins): rename help plugin
