+++ src/plugins/help/command.ts
